@@ -2,6 +2,7 @@
   "name": "city-foundation-next",
   "version": "1.1.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev --turbopack",
     "develop": "next dev",
@@ -15,20 +16,25 @@
   },
   "dependencies": {
     "@js-joda/core": "^5.6.4",
+    "@payloadcms/db-postgres": "^3.33.0",
+    "@payloadcms/next": "^3.33.0",
+    "@payloadcms/payload-cloud": "^3.33.0",
+    "@payloadcms/richtext-lexical": "^3.33.0",
     "@svgr/webpack": "8.1.0",
     "classnames": "^2.3.2",
     "graphql": "^16.8.1",
-    "graphql-request": "^4.3.0",
-    "graphql-tag": "^2.12.6",
+    "graphql-request": "6.1.0",
+    "graphql-tag": "2.12.6",
     "lodash": "^4.17.21",
     "mailgun-js": "^0.22.0",
     "next": "15.2.4",
     "next-i18next": "^11.3.0",
     "next-plugin-custom-babel-config": "^1.0.5",
     "next-transpile-modules": "^8.0.0",
+    "payload": "^3.33.0",
     "postcss-import": "^14.1.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-i18next": "^11.18.6",
     "react-markdown": "10.1.0",
     "react-select": "^5.10.1",
@@ -38,13 +44,13 @@
     "tailwindcss": "3.4.17"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^2.16.5",
-    "@graphql-codegen/typescript": "^2.8.8",
-    "@graphql-codegen/typescript-graphql-request": "^4.5.9",
-    "@graphql-codegen/typescript-operations": "^2.5.13",
+    "@graphql-codegen/cli": "^5.0.5",
+    "@graphql-codegen/typescript": "^4.1.5",
+    "@graphql-codegen/typescript-graphql-request": "^6.2.0",
+    "@graphql-codegen/typescript-operations": "^4.5.1",
     "@next/eslint-plugin-next": "15.2.4",
-    "@types/react": "19.0.12",
-    "@types/react-dom": "19.0.4",
+    "@types/react": "19.1.0",
+    "@types/react-dom": "19.1.1",
     "@types/react-select": "^5.0.1",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
@@ -92,7 +98,7 @@
     "npm": "10.9.2"
   },
   "resolutions": {
-    "@types/react": "19.0.12",
-    "@types/react-dom": "19.0.4"
+    "@types/react": "19.1.0",
+    "@types/react-dom": "19.1.1"
   }
 }
