@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   trailingComma: 'all',
   tabWidth: 2,
   semi: false,
