@@ -1,6 +1,4 @@
-const path = require('path')
-
-module.exports = {
+export default {
   i18n: {
     defaultLocale: 'sk',
     locales: ['en', 'sk'],
@@ -8,5 +6,4 @@ module.exports = {
     localeDetection: false,
   },
   reloadOnPrerender: process.env.NODE_ENV === 'development',
-  localePath: path.resolve('./public/locales'),
 }
