@@ -1,10 +1,6 @@
-const { join } = require('path')
-
-module.exports = {
+export default {
   plugins: {
-    tailwindcss: {
-      config: join(__dirname, 'tailwind.config.js'),
-    },
+    tailwindcss: {},
     autoprefixer: {},
   },
 }
